property writes are def nodes
